instatwo/backend/requirements.txt
asgiref==3.7.2
Django==3.2.23
django-cors-headers==4.3.1
djangorestframework==3.14.0
pillow==10.2.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4